(Tech) split Home Header into separate component
