+++ src/features/home/components/index.ts
@@ -1,4 +1,5 @@
 export { BusinessModule } from './BusinessModule'
 export { ExclusivityModule } from './ExclusivityModule'
 export { HomeBodyPlaceholder } from './HomeBodyPlaceholder'
+export { HomeHeader } from './HomeHeader'
 export { OffersModule } from './OffersModule'
